+++ src/sources/kubernetes_logs/transform_utils/optional.rs
