+++ src/transforms/util/mod.rs
@@ -1,8 +1,2 @@
 #[cfg(any(feature = "transforms-lua"))]
 pub mod runtime_transform;
-
-// TODO: make these part of the vector core - enabled unconditionally.
-#[cfg(any(feature = "sources-kubernetes-logs"))]
-pub mod optional;
-#[cfg(any(feature = "sources-kubernetes-logs"))]
-pub mod pick;
